+++ eyes.selenium.java/src/test/java/com/applitools/eyes/renderingGrid/TestRenderingGridMultiThreadService.java
@@ -10,7 +10,7 @@ import com.applitools.eyes.selenium.Eyes;
 import com.applitools.eyes.selenium.fluent.Target;
 import com.applitools.eyes.utils.SeleniumUtils;
 import com.applitools.eyes.utils.TestUtils;
-import com.applitools.eyes.visualgrid.model.RenderBrowserInfo;
+import com.applitools.eyes.visualgrid.model.DesktopBrowserInfo;
 import com.applitools.eyes.visualgrid.services.VisualGridRunner;
 import com.applitools.utils.GeneralUtils;
 import org.openqa.selenium.WebDriver;
@@ -49,9 +49,9 @@ public final class TestRenderingGridMultiThreadService {
             public void run() {
 
                 TestThreadMethod("VG-2ThreadBatchC11",
-                        new RenderBrowserInfo(800, 600, BrowserType.CHROME, baselineEnvName),
-                        new RenderBrowserInfo(700, 500, BrowserType.CHROME, baselineEnvName),
-                        new RenderBrowserInfo(400, 300, BrowserType.CHROME, baselineEnvName));
+                        new DesktopBrowserInfo(800, 600, BrowserType.CHROME, baselineEnvName),
+                        new DesktopBrowserInfo(700, 500, BrowserType.CHROME, baselineEnvName),
+                        new DesktopBrowserInfo(400, 300, BrowserType.CHROME, baselineEnvName));
             }
         });
 
@@ -59,8 +59,8 @@ public final class TestRenderingGridMultiThreadService {
             @Override
             public void run() {
                 TestThreadMethod("VG-2ThreadBatchC22",
-                        new RenderBrowserInfo(840, 680, BrowserType.CHROME, baselineEnvName),
-                        new RenderBrowserInfo(750, 530, BrowserType.CHROME, baselineEnvName));
+                        new DesktopBrowserInfo(840, 680, BrowserType.CHROME, baselineEnvName),
+                        new DesktopBrowserInfo(750, 530, BrowserType.CHROME, baselineEnvName));
             }
         });
 
@@ -75,7 +75,7 @@ public final class TestRenderingGridMultiThreadService {
         }
     }
 
-    private void TestThreadMethod(String batchName, RenderBrowserInfo... browsersInfo) {
+    private void TestThreadMethod(String batchName, DesktopBrowserInfo... browsersInfo) {
         Eyes eyes = new Eyes(renderingManager);
         eyes.setBatch(new BatchInfo(batchName));
         Configuration configuration = new Configuration();