+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [vNext]
+## [3.165.0]
 ### Added
 - Supported rendering on ios simulators. [Trello 1872](https://trello.com/c/bykk2rzB)
 - Added support for Visual Locators. [Trello 1788](https://trello.com/c/dEeEDiIY)